refactor: share conditional token balance helpers
